+++ lib/src/widgets/index.dart
@@ -1,1 +1,2 @@
-export 'rainbow_circle_toy_loading_indicator.dart';
+export 'chakra_loading_indicator.dart';
+export 'slinky_loading_indicator.dart';