+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [2.0.0] - 03/06/2022
+
+* [Breaking changes] New implementation for Turtle Graphics.
+* Example for making a fractal art by using TurtleGraphicsPainter.
+* Create ChakraLoadingIndicator.
+* Create SlinkyLoadingIndicator.
+
 ## [1.1.0] - 25/05/2022
 
 * Migrate to flutter 3.0.1.